refactor(core): #HttpRequestBus
