+++ packages/core/src/http/server/http.server.tokens.ts
@@ -1,6 +1,4 @@
-import { Subject } from 'rxjs';
 import { createContextToken } from '../../context/context.token.factory';
-import { HttpServer, HttpRequest } from '../http.interface';
+import { HttpServer } from '../http.interface';
 
 export const HttpServerClientToken = createContextToken<HttpServer>('HttpServerClientToken');
-export const HttpRequestBusToken = createContextToken<Subject<HttpRequest>>('HttpRequestBusToken');
